requirements.txt
streamlit>=1.36
pillow>=10.0
pillow-avif-plugin>=1.4
